# docs/generated/seller-price-monitoring-openapi.yaml
openapi: 3.1.0
info:
  title: Seller Price Monitoring API
  version: 0.1.0
  description: >
    Draft OpenAPI specification for the seller price monitoring service.
    This file mirrors the human-readable API design docs and keeps request/response
    shapes aligned with the schema and wireframes.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      description: |-
        401: email or password is invalid.
        403: account is suspended or login is blocked.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        <<: *AuthErrorResponses
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an access token
      description: |-
        401: refresh token is missing, expired, or invalid.
        403: the session has been revoked.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        <<: *AuthErrorResponses
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current session
      description: |-
        401: access token is missing or expired.
        403: the current session is already revoked.
      responses:
        <<: *AuthErrorResponses
        '204':
          description: Logged out
  /me:
    get:
      tags: [Auth]
      summary: Current user and accessible organizations
      description: |-
        401: access token is missing or expired.
        403: the account is suspended or permissions were revoked.
      responses:
        <<: *ReadErrorResponses
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      example:
        error:
          code: VALIDATION_ERROR
          message: ownerProductUrl is required
          details:
            ownerProductUrl: [required]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    LoginRequest:
      type: object
      required: [email, password]
      example:
        email: seller@example.com
        password: password123
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RefreshRequest:
      type: object
      required: [refreshToken]
      example:
        refreshToken: jwt-refresh-token
      properties:
        refreshToken:
          type: string
    UserSummary:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
    Organization:
      type: object
      required: [id, name, plan, status, timezone, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        plan:
          type: string
          enum: [starter, growth, agency]
        status:
          type: string
          enum: [active, suspended, closed]
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthTokenResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [accessToken, refreshToken, user]
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            user:
              $ref: '#/components/schemas/UserSummary'
    LoginResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [accessToken, refreshToken, user]
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            user:
              $ref: '#/components/schemas/UserSummary'
    MeResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [user, organizations]
          properties:
            user:
              $ref: '#/components/schemas/UserSummary'
            organizations:
              type: array
              items:
                $ref: '#/components/schemas/Organization'
...