# src/main/java/kr/co/harness/spm/auth/dto/LoginRequest.java
package kr.co.harness.spm.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @Email @NotBlank String email,
        @NotBlank String password
) {
}